# LocalCudaWorkerService.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LocalCudaWorkerService API",
    "description": "ASP.NET REST-API using CUDA as an online worker for executing kernels via HTTP.",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "github: alarmclock-kisser",
      "email": "marcel.king91299@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Worker/info": {
      "get": {
        "tags": [
          "Worker"
        ],
        "operationId": "GetInfo",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiConfiguration"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiConfiguration"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiConfiguration"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Worker/config": {
      "get": {
        "tags": [
          "Worker"
        ],
        "operationId": "GetConfig",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiConfiguration"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiConfiguration"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiConfiguration"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Worker/connect-to-server": {
      "get": {
        "tags": [
          "Worker"
        ],
        "parameters": [
          {
            "name": "overwriteServerUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Worker/unregister": {
      "delete": {
        "tags": [
          "Worker"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Worker/api-log": {
      "get": {
        "tags": [
          "Worker"
        ],
        "parameters": [
          {
            "name": "maxEntries",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiConfiguration": {
        "type": "object",
        "properties": {
          "applicationName": {
            "type": "string",
            "nullable": true
          },
          "localServerIp": {
            "type": "string",
            "nullable": true
          },
          "localServerUrl": {
            "type": "string",
            "nullable": true
          },
          "localServerPort": {
            "type": "integer",
            "format": "int32"
          },
          "useHttps": {
            "type": "boolean"
          },
          "externalServerAddress": {
            "type": "string",
            "nullable": true
          },
          "successfullyConnectedToExternalServer": {
            "type": "boolean"
          },
          "registeredAtExternalServer": {
            "type": "boolean"
          },
          "maxUploadSizeMb": {
            "type": "integer",
            "format": "int32"
          },
          "defaultDeviceIndex": {
            "type": "integer",
            "format": "int32"
          },
          "defaultDeviceName": {
            "type": "string",
            "nullable": true
          },
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}